+++ src/schemas/user_institution_schema.py
@@ -1,0 +1,19 @@
+from pydantic import BaseModel
+
+from src.schemas.profile_schemas import ProfileGet
+from src.schemas.institutions_schema import InstitutionGet
+from src.schemas.users_schemas import UsersLoginResponse
+
+
+class UserInstitutionSchema(BaseModel):
+    """Schema for user institution association."""
+
+    user: UsersLoginResponse
+    institution: InstitutionGet
+    profile: ProfileGet
+
+    class Config:
+        """Configure Pydantic to allow population from ORM objects and provide an example."""
+
+        from_attributes = True
+        json_schema_extra = {"example": {"token": "token_jwt_aqui"}}